% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{b}
\alias{b}
\title{maturity adjustment factor b}
\usage{
b(PD)
}
\arguments{
\item{PD}{enter the PD as a decimal number}
}
\value{
A Number
}
\description{
The maturity adjustment factor b according to Sub-Section 2 Article 153 1. iii)
}
\examples{

PD <- 0.002
b(PD)


}
