% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_retail.R
\name{rw_retail}
\alias{rw_retail}
\title{Risk weight for non defaulted retail exposures}
\usage{
rw_retail(PD, LGD, R = 0.15, SK = 1.06)
}
\arguments{
\item{PD}{enter the PD as a decimal number}

\item{LGD}{enter the LGD as a decimal number}

\item{R}{R is set to 0.15 for retail exposures secured by immovable property according to
to Sub-Section 2 Article 154 3.}

\item{SK}{default is 1.06 according to the basel formula. You can change the value to 1, for example, to simulatethe effects
from basel IV}
}
\value{
A Number
}
\description{
Calculated the risk weight according to Sub-Section 2 Article 154 1. ii) for non defaulted retail exposures (0 < PD < 1).
In the actual version, the correlation factor R is set to 0.15 for retail exposures secured by immovable property according to
to Sub-Section 2 Article 154 3..
}
\examples{

PD <- 0.002
LGD <- 0.12
rw_retail(PD,LGD)

#simulate Basel IV effects without factor 1.06#'
PD <- 0.002
LGD <- 0.12
rw_retail(PD,LGD, SK=1)

}
