% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_R.R
\name{corr_R}
\alias{corr_R}
\title{coefficient of correlation R}
\usage{
corr_R(PD)
}
\arguments{
\item{PD}{enter the PD as a decimal number}
}
\value{
A Number
}
\description{
The coefficient of correlation R according to Sub-Section 2 Article 153 1. iii)
}
\examples{

PD <- 0.002
corr_R(PD)


}
